Update contrast to use cookies
This is an updated version of toggle-contrast.js that uses cookies as the primary storage, with localStorage as a fallback.

This is to fix the domain issue between Glossary and the rest of WFU.

Added as a new file to ensure no break in functionality for current users as well to test on staging without affecting production.